Add screenshots to report
Adds screenshots of the elements that failed the color contrast test
to the output HTML report. The screenshots are sorted by domain
and correspond to a page title.
Note: the filenames remain in UTF8 format for easier correlation with
the titles, this could cause errors in some situations where title
are not valid filenames.